Refactor configuration and hook signing
