+++ lib/overcommit.rb
@@ -5,6 +5,7 @@ require 'overcommit/utils/file_utils'
 require 'overcommit/utils'
 require 'overcommit/git_version'
 require 'overcommit/configuration_validator'
+require 'overcommit/signer'
 require 'overcommit/configuration'
 require 'overcommit/configuration_loader'
 require 'overcommit/hook/base'
