+++ lib/overcommit/hook_signer.rb
@@ -1,8 +1,6 @@
-require 'fileutils'
-
 module Overcommit
   # Calculates, stores, and retrieves stored signatures of hook plugins.
-  class HookSigner # rubocop:disable Metrics/ClassLength
+  class HookSigner < Overcommit::Signer
     attr_reader :hook_name
 
     # We don't want to include the skip setting as it is set by Overcommit
@@ -16,6 +14,8 @@ module Overcommit
       @hook_name = hook_name
       @config = config
       @context = context
+
+      @key = signature_config_key
     end
 
     # Returns the path of the file that should be incorporated into this hooks
@@ -55,90 +55,7 @@ module Overcommit
         Overcommit::GitRepo.tracked?(file)
     end
 
-    # Return whether the signature for this hook has changed since it was last
-    # calculated.
-    #
-    # @return [true,false]
-    def signature_changed?
-      changed = signature != stored_signature
-
-      if changed && has_history_file
-        changed = !signature_in_history_file(signature)
-      end
-
-      changed
-    end
-
-    # Update the current stored signature for this hook.
-    def update_signature!
-      updated_signature = signature
-
-      result = Overcommit::Utils.execute(
-        %w[git config --local] + [signature_config_key, updated_signature]
-      )
-
-      unless result.success?
-        raise Overcommit::Exceptions::GitConfigError,
-              "Unable to write to local repo git config: #{result.stderr}"
-      end
-
-      add_signature_to_history(updated_signature)
-    end
-
     private
-
-    def add_signature_to_history(signature)
-      # Now we must update the history file with the new signature
-      # We want to put the newest signatures at the top, since they are more
-      # likely to be used, and will be read sooner
-      signatures = []
-      if has_history_file
-        signatures = (File.readlines history_file).first(@config.signature_history - 1)
-      else
-        FileUtils.mkdir_p(File.dirname(history_file))
-      end
-
-      File.open(history_file, 'w') do |fh|
-        fh.write("#{signature}\n")
-        signatures.each do |old_signature|
-          fh.write(old_signature)
-        end
-      end
-    end
-
-    def signature_in_history_file(signature)
-      unless has_history_file
-        return false
-      end
-
-      found = false
-      File.open(history_file, 'r') do |fh|
-        # Process the header
-        until (line = fh.gets).nil?
-          line.chomp
-
-          if line == signature
-            found = true
-            break
-          end
-        end
-      end
-
-      found
-    end
-
-    # Does the history file exist
-    def has_history_file
-      STDERR.puts 'checking history file'
-      File.exist?(history_file)
-    end
-
-    # Determine the absolute path for this signer's history file
-    def history_file
-      File.join(@config.hook_signature_directory,
-                @context.hook_type_name,
-                "#{Overcommit::Utils.snake_case(@hook_name)}.rb")
-    end
 
     # Calculates a hash of a hook using a combination of its configuration and
     # file contents.
@@ -155,21 +72,6 @@ module Overcommit
 
     def hook_contents
       File.read(hook_path)
-    end
-
-    def stored_signature
-      result = Overcommit::Utils.execute(
-        %w[git config --local --get] + [signature_config_key]
-      )
-
-      if result.status == 1 # Key doesn't exist
-        return ''
-      elsif result.status != 0
-        raise Overcommit::Exceptions::GitConfigError,
-              "Unable to read from local repo git config: #{result.stderr}"
-      end
-
-      result.stdout.chomp
     end
 
     def signature_config_key
